Code to use the nu.xom.Element linkedobject instead of Node Object to create the Tree structure.
Element is now used as input to getMapStructure function to generate the Json Map and Json structure.

it has removed the need to wrote the xml parse of our own.

Also we have introduced the XMLTOMapConversionPreProcessor.java to handle the special scenarios before passing the Element object to map structre function